Rename and rewrite test
Simplify complicated test that seems to use overlapping hosts
between applications.
Rewrite to just check that subscribing to supermodel config works
and adding a service to the application is reflected in config
