+++ tests/cloudconfig/supermodel/supermodelbundle/manifest.mf
